ltc: move all bignum functions to mpa_desc.c
Moves all crypto_bignum*() functions from tee_ltc_provider.c to
mpa_desc.c in order to contain libmpa dependencies.
